Adding implementation in method getWalletSyncData

+++ src/main/java/org/kryptokrona/sdk/model/http/WalletSyncData.java
@@ -16,7 +16,7 @@ import java.util.ArrayList;
 @NoArgsConstructor
 public class WalletSyncData {
 
-    private long blockCount;
+    private double blockCount;
 
     private ArrayList<String> checkpoints;
